requirements.txt
cliff
datakit-core
PyGithub
